[sui-proxy/ requested comments] Summary:
* addressing comments from https://github.com/MystenLabs/sui/pull/9123

Test Plan:

cargo build/test/clippy
