+++ crates/sui-proxy/src/config.rs
@@ -18,7 +18,7 @@ pub struct ProxyConfig {
 }
 
 #[serde_as]
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Deserialize, Serialize, Default)]
 #[serde(rename_all = "kebab-case")]
 pub struct RemoteWriteConfig {
     // TODO upgrade to https
@@ -28,6 +28,11 @@ pub struct RemoteWriteConfig {
     /// username is used for posting data to the remote_write api
     pub username: String,
     pub password: String,
+
+    /// Sets the maximum idle connection per host allowed in the pool.
+    /// <https://docs.rs/reqwest/latest/reqwest/struct.ClientBuilder.html#method.pool_max_idle_per_host>
+    #[serde(default = "pool_max_idle_per_host_default")]
+    pub pool_max_idle_per_host: usize,
 }
 
 #[serde_as]
@@ -51,14 +56,22 @@ pub struct PeerValidationConfig {
     pub private_key: Option<String>,
 }
 
+/// the default idle worker per host (reqwest to remote write url call)
+fn pool_max_idle_per_host_default() -> usize {
+    8
+}
+
+/// the default hostname we will use if not provided
 fn hostname_default() -> Option<String> {
     Some("localhost".to_string())
 }
 
+/// the default remote write url
 fn remote_write_url() -> String {
     "http://metrics-gw.testnet.sui.io/api/v1/push".to_string()
 }
 
+/// load our config file from a path
 pub fn load<P: AsRef<std::path::Path>, T: DeserializeOwned + Serialize>(path: P) -> Result<T> {
     let path = path.as_ref();
     debug!("Reading config from {:?}", path);
