+++ crates/sui-proxy/src/main.rs
@@ -79,23 +79,3 @@ const GIT_REVISION: &str = {
     }
 };
 const VERSION: &str = const_str::concat!(env!("CARGO_PKG_VERSION"), "-", GIT_REVISION);
-
-/// var extracts environment variables at runtime with a default fallback value
-/// if a default is not provided, the value is simply an empty string if not found
-/// This function will return the provided default if env::var cannot find the key
-/// or if the key is somehow malformed.
-#[macro_export]
-macro_rules! var {
-    ($key:expr) => {
-        match std::env::var($key) {
-            Ok(val) => val,
-            Err(_) => "".into(),
-        }
-    };
-    ($key:expr, $default:expr) => {
-        match std::env::var($key) {
-            Ok(val) => val,
-            Err(_) => $default,
-        }
-    };
-}
